Refactor common styling to DETab and DETabs
